Add script to automate s3 build and publish process
A large part of the iOS SDK deploy process is creating the various
binary flavors, uploading them to s3, and then testing that those
uploads worked by downloading them -- those binaries are used in the
next step of the process when they are attached to the github release.

This script automates that build, upload, and download process and
can be run after `mbx auth`.